create review controller

+++ controllers/categoryController.js
@@ -1,7 +1,6 @@
 const catchAsyncErrors = require("../middleware/catchAsyncErrors");
 const Category = require("../models/categoryModel");
 const ErrorHandler = require("../utils/errorhandler");
-const cloudinary = require("cloudinary");
 
 exports.addCategory = catchAsyncErrors(async (req, res, next) => {
   const { categoryName } = req.body;
@@ -15,23 +14,22 @@ exports.addCategory = catchAsyncErrors(async (req, res, next) => {
 
   res.status(201).json({
     success: true,
-    message: "Category Added!!!",
+    message: "Category created",
   });
 });
 
-exports.getAllCategories = catchAsyncErrors(async (req, res) => {
+exports.getAllCategories = catchAsyncErrors(async (req, res, next) => {
   const allCategories = await Category.find();
-
-  res.status(201).json({
+  res.status(200).json({
     success: true,
     allCategories,
   });
 });
 
-exports.deleteCategory = catchAsyncErrors(async (req, res) => {
-  const categoryId = await Category.findById(req.params.id);
+exports.deleteCategory = catchAsyncErrors(async (req, res, next) => {
+  const categoryId = await Category.findById(req.params.categoryId);
   const category = await Category.findById(categoryId);
-  console.log(category);
+  if (!category) return next(new ErrorHandler("cateogry not found", 400));
   await Category.deleteOne({ _id: categoryId });
   res.status(200).json({
     success: true,
